rename checkpoint ui components
