+++ src/entities/checkpoint/ui/checkpoint-group-card/CheckpointGroupCard.tsx
@@ -1,25 +1,25 @@
 import type { ComponentPropsWithoutRef, ReactElement } from 'react'
-import styles from './CheckpointGroup.module.css'
+import styles from './CheckpointGroupCard.module.css'
 import type { CheckpointGroup } from '../../model/types'
-import { Checkpoint } from '../checkpoint/Checkpoint'
+import { CheckpointRow } from '../checkpoint-row/CheckpointRow'
 
-interface CheckpointGroupProps extends ComponentPropsWithoutRef<'div'> {
+interface CheckpointGroupCardProps extends ComponentPropsWithoutRef<'div'> {
   group: CheckpointGroup
   actions?: ReactElement
 }
 
-export function CheckpointGroup({ group, actions, children, className, ...props }: CheckpointGroupProps) {
+export function CheckpointGroupCard({ group, actions, children, className, ...props }: CheckpointGroupCardProps) {
   const renderCheckpoints = () =>
     group.checkpoints.map(c => (
       <li key={c.title}>
-        <Checkpoint checkpoint={c} />
+        <CheckpointRow checkpoint={c} />
       </li>
     )) // TODO проверить, валидируется ли на бэкенде
 
   return (
     <div className={`${styles.container} ${className ?? ''}`} {...props}>
       <div className={styles.header}>
-        <h5>{group.title}</h5>
+        <h5 className={styles.title}>{group.title}</h5>
         <div className={styles.actions}>{actions}</div>
       </div>
       <ul className={styles.list}>{renderCheckpoints()}</ul>
