+++ src/entities/checkpoint/ui/checkpoint-group-card/CheckpointGroupCardSkeleton.module.css
@@ -1,5 +1,5 @@
 .container {
-	width: 400px;
+	width: 330px;
 }
 
 .title {
