+++ src/entities/checkpoint/ui/checkpoint-group-card/CheckpointGroupCardSkeleton.tsx
@@ -1,8 +1,8 @@
-import styles from './CheckpointGroup.module.css'
-import s from './CheckpointGroupSkeleton.module.css'
+import styles from './CheckpointGroupCard.module.css'
+import s from './CheckpointGroupCardSkeleton.module.css'
 import { TextSkeleton } from '@/shared'
 
-export function CheckpointGroupSkeleton() {
+export function CheckpointGroupCardSkeleton() {
   return (
     <div className={`${styles.container} ${s.container}`}>
       <div className={styles.header}>
@@ -10,7 +10,7 @@ export function CheckpointGroupSkeleton() {
         <TextSkeleton className={s.actions} />
       </div>
       <div className={styles.list}>
-        <TextSkeleton rows={3} />
+        <TextSkeleton rows={4} />
       </div>
     </div>
   )
