--- src/entities/checkpoint/ui/checkpoint-group/CheckpointGroup.module.css
@@ -1,31 +1,0 @@
-.container {
-	width: max-content;
-
-	display: flex;
-	flex-direction: column;
-	gap: 10px;
-
-	padding: 20px;
-	border-radius: 10px;
-	
-	background-color: var(--sec-color);
-}
-
-.header {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	gap: 7px;
-}
-
-.actions {
-	display: flex;
-	align-items: center;
-	gap: 5px;
-}
-
-.list {
-	display: flex;
-	flex-direction: column;
-	gap: 7px;
-}
